Less CloudFormation [broken]
The prep-stack was constantly failing with an opaque 400 error, and,
since it was only there to create a versioned S3 bucket, I decided to
use the AWS CLI instead.

+ Make use of the generic AWS CLI environment variables instead of
passing --region and --profile arguments to every command.

+ Add SecurityGroupIds parameter to the main stack.

The main stack is not deployable now because of incorrect VpcConfig for
TestLambdaFunction.