one express to serve them all
# backend: Introduced ExpressAdapter

Express turned out to be handy for serving the frontend files which,
with a relatively small routing logic, frees me from creating the
intricate directory structure, create HTML pages from the template and
move Node modules around. Now all that logic lives in
backend/src/App/ExpressAdapter.ts.

# Consolidated the Makefiles (almost)

Trying to streamline the workflow so that I don’t need to cd back and
forth between backend and frontend. There are a couple more pieces in
the backend/Makefile that I will return to later.

# backend: Turned on esModuleInterop

This allows for this:

  import fs from "fs"

instead of this:

  import * as fs from "fs"

I prefer the former because it’s more concise, and because I don’t like
to type `* as`.

# backend: Added the debug module

It seems handy.

# backend: no emit

I’m thinking that not compiling the backend code will streamline the dev
workflow: this frees me from needing to compile the backend code besides
restarting it, and also the stack-traces are clearer.

# frontend: no minification so far

I will probably bring it back later, for production; in dev it seems
like I don’t need it for now.