Validation: predicates return error codes
...instead of humanly-readable error messages.

Why? -- Because I only need the humanly-readable messages specifically
on the front-end, when I validate the input before sending it to the
back-end. Back-end doesn’t really need the humanly-readable messages, it
only needs to precisely signal the issue.

This way different front-end clients can have the exact messages that
they need for their particular context.