Don’t use SameSite for session cookie
Cookies without a SameSite attribute are as SameSite=Lax.

See https://web.dev/samesite-cookies-explained/.

I’m making this change to fix the issue where the session would be lost
when accessing the reset-password or reset-email links in Gmail. In
short, when the request initiator, in this case Gmail, is on a different
site than the destination address, cookies with SameSite=Strict are not
sent, which leads to express-session concluding that there is no session
and creating a new one.