Introduce UserSession2
I’m thinking I could have a userSession on the Express.js session, and
then the app would work with that.

The userSession would be of type UserSession and would have userId and
profile properties. I already have a UserSession type, which is
different, and if I were to change it in place it would be a huge
change, so I will introduce a new type UserSession2 that, and migrate
all the code to use that instead of the existing one, and then will
remove the current UserSession, and rename UserSession2 to UserSession.

Before log-in userSession would be undefined. The Login scenario handler
would populate it.